builder: move file download to helpers
Download a file is a generic operation. This patch moves this
functionality from builder to helpers to allow it to be reused
by other packages.
